requirements.txt
dnspython
tabulate
pyyaml